requirements.txt
cssselect
lemminflect
lxml
pymongo
PyYAML
Scrapy